+++ packages/data-stores/src/plans/queries/use-priced-api-site-plans.ts
@@ -3,10 +3,6 @@ import { useCallback } from '@wordpress/element';
 import wpcomRequest from 'wpcom-proxy-request';
 import type { PricedAPISitePlan } from '../types';
 
-/*
- * - Plans from `/sites/[siteId]/plans`, unlike `/plans`, are returned indexed by product_id, and do not inlcude that in the plan's payload.
- * - UI works with product/plan slugs everywhere, so returned index is transformed to be keyed by product_slug
- */
 interface PricedAPISitePlansIndex {
 	[ planSlug: string ]: PricedAPISitePlan;
 }
@@ -18,6 +14,10 @@ interface Props {
 	siteId?: string | number | null;
 }
 
+/**
+ * - Plans from `/sites/[siteId]/plans`, unlike `/plans`, are returned indexed by product_id, and do not inlcude that in the plan's payload.
+ * - UI works with product/plan slugs everywhere, so returned index is transformed to be keyed by product_slug
+ */
 function usePricedAPISitePlans( { siteId }: Props ) {
 	return useQuery< PricedAPISitePlansIndexServer, Error, PricedAPISitePlansIndex >( {
 		queryKey: [ 'site-plans', siteId ],
@@ -27,7 +27,7 @@ function usePricedAPISitePlans( { siteId }: Props ) {
 				apiVersion: '1.3',
 			} ),
 		select: useCallback( ( data: PricedAPISitePlansIndexServer ) => {
-			return Object.keys( data ).reduce( ( acc, productId ) => {
+			return Object.keys( data ).reduce< PricedAPISitePlansIndex >( ( acc, productId ) => {
 				const plan = data[ Number( productId ) ];
 				return {
 					...acc,
