+++ packages/data-stores/src/products-list/types.ts
@@ -5,17 +5,6 @@ export interface Dispatch {
 	dispatch: DispatchFromMap< typeof actions >;
 }
 
-export type IntroductoryOfferTimeUnit = 'day' | 'week' | 'month' | 'year';
-
-export interface ProductIntroductoryOffer {
-	cost_per_interval: number;
-	interval_count: number;
-	interval_unit: IntroductoryOfferTimeUnit;
-	should_prorate_when_offer_ends: boolean;
-	transition_after_renewal_count: number;
-	usage_limit: number | null;
-}
-
 export interface ProductsListItem {
 	available: boolean;
 	combined_cost_display: string;
